+++ src/Form/ContractByUserForm.php
@@ -9,40 +9,15 @@
 
 namespace App\Form;
 
-use App\Form\Type\UserType;
-use App\Form\Type\YearPickerType;
-use App\Reporting\YearByUser\YearByUser;
 use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\OptionsResolver\OptionsResolver;
 
 /**
  * @internal
  */
 final class ContractByUserForm extends AbstractType
 {
-    public function buildForm(FormBuilderInterface $builder, array $options): void
+    public function getParent()
     {
-        $builder->add('date', YearPickerType::class, [
-            'model_timezone' => $options['timezone'],
-            'view_timezone' => $options['timezone'],
-            'start_date' => $options['start_date'],
-        ]);
-
-        if ($options['include_user']) {
-            $builder->add('user', UserType::class, ['width' => false]);
-        }
-    }
-
-    public function configureOptions(OptionsResolver $resolver): void
-    {
-        $resolver->setDefaults([
-            'data_class' => YearByUser::class,
-            'timezone' => date_default_timezone_get(),
-            'start_date' => new \DateTime(),
-            'include_user' => false,
-            'csrf_protection' => false,
-            'method' => 'GET',
-        ]);
+        return YearByUserForm::class;
     }
 }
